feat(instant_charge): Add API route to show a fee
