+++ lago_python_client/models/credit.py
@@ -1,12 +1,7 @@
 from pydantic import BaseModel, Field
 from typing import Optional, List
 
-
-class InvoiceItemResponse(BaseModel):
-    lago_id: Optional[str]
-    type: Optional[str]
-    code: Optional[str]
-    name: Optional[str]
+from .invoice_item import InvoiceItemResponse
 
 
 class InvoiceShortDetails(BaseModel):
