+++ tests/fixtures/fee.json
@@ -1,0 +1,17 @@
+{
+  "fee": {
+    "lago_id": "5eb02857-a71e-4ea2-bcf9-57d3a41bc6ba",
+    "lago_group_id": "1a2d9c8d-5875-4688-9854-5ccfd414bc5e",
+    "item": {
+      "type": "charge",
+      "code": "fee_code",
+      "name": "Fee Code"
+    },
+    "amount_cents": 120,
+    "amount_currency": "EUR",
+    "vat_amount_cents": 20,
+    "vat_amount_currency": "EUR",
+    "units": "10.0",
+    "events_count": 10
+  }
+}
